back/requirements.txt
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.4.0
Django==5.1.3
django-environ==0.11.2
django-filter==24.3
djangorestframework==3.15.2
djangorestframework-gis==1.1
idna==3.10
mysqlclient==2.2.6
requests==2.32.3
sqlparse==0.5.1
tzdata==2024.2
urllib3==2.2.3
